+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Upgrade dependencies for OpenCV 4.2.0, FFmpeg 4.2.2, and Tesseract 4.1.1
  * Add `Seekable` and `SeekableByteArrayOutputStream` to be used with `FFmpegFrameRecorder` ([pull #1350](https://github.com/bytedeco/javacv/pull/1350))
  * Update `RealSense2FrameGrabber` with support for sensor options and fix for multiple devices ([pull #1348](https://github.com/bytedeco/javacv/pull/1348))